# validateai/public/swagger.yaml
openapi: 3.0.3
info:
  title: Bralidus API Platform
  version: "2.0.0"
  description: |
    Plataforma de APIs de Inteligencia de Mercado, Mercado Público (B2G), Datos Macroeconómicos, Grafo Societario y GraphRAG como Servicio (RaaS).

    Centraliza datos de Mercado Público (Licitaciones, Compra Ágil, Órdenes de Compra, Organismos, Proveedores, Benchmarks, Métricas M1-M10), indicadores macroeconómicos (BCCh, FRED, yfinance, CMF, INE, SEIA, Diario Oficial), Grafo Societario S-Pulse y motor vectorial pgvector.

    ## Autenticación
    Todas las rutas bajo `/api/v1` requieren una API Key válida en la cabecera HTTP:
    ```
    Authorization: Bearer val_live_XXXX...
    ```
    Genera y gestiona tus llaves de acceso en el [Developer Portal](https://bralidus.scouttech.lat/developers).

    ## Rate Limiting & Créditos
    - Cada endpoint consume un monto ponderado de créditos en función de su complejidad.
    - Headers de respuesta estándar: `X-RateLimit-Limit-Credits`, `X-RateLimit-Remaining-Credits`, `X-RateLimit-Tier`, `X-RateLimit-Request-Cost`.

  contact:
    name: Bralidus API Team
    url: https://bralidus.scouttech.lat
  license:
    name: Propietario
servers:
  - url: https://fcdhcntyvsydnvjwopfe.supabase.co/functions/v1/api-v1
    description: Producción Edge Gateway

tags:
  - name: Mercado Público (B2G)
    description: Licitaciones, Compra Ágil, Órdenes de Compra, Organismos Compradores, Proveedores, Benchmarks y Métricas M1-M10
  - name: Datos Económicos & Macro
    description: Indicadores macroeconómicos locales y globales (BCCh, FRED, yfinance, CMF Insolvencias, INE Empleo, SEIA, Diario Oficial)
  - name: Bralidus Intelligence (GraphRAG)
    description: Búsqueda semántica fusionada con datos macroeconómicos, normativos y Mixture of Experts (MoE)
  - name: RAG & Vault Vectorial
    description: Búsqueda vectorial semántica pgvector e ingesta de documentos y Vault privado
  - name: Grafo Societario (S-Pulse)
    description: Redes societarias, mallas de empresas, socios, directivas y trazabilidad legal auditable
  - name: Webhooks & Servicios
    description: Notificaciones asíncronas en tiempo real, validación de startups y health check

paths:
  # ── Mercado Público: Licitaciones & Compra Ágil ───────────────────────────
  /api/v1/mercado-publico/health:
    get:
      tags: [Mercado Público (B2G)]
      summary: Estado operativo de servicios Mercado Público
      responses:
        '200':
          description: Estado del servicio.

  /api/v1/mercado-publico/opportunities:
    get:
      tags: [Mercado Público (B2G)]
      summary: Buscador unificado (licitaciones y compras ágiles)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [tender, agile_purchase]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Listado de oportunidades.

  /api/v1/mercado-publico/licitaciones:
    get:
      tags: [Mercado Público (B2G)]
      summary: Buscador de licitaciones públicas
      parameters:
        - name: fecha_inicio
          in: query
          schema:
            type: string
        - name: fecha_fin
          in: query
          schema:
            type: string
        - name: estado
          in: query
          schema:
            type: string
        - name: codigo_organismo
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Listado paginado de licitaciones.

  /api/v1/mercado-publico/licitaciones/{codigo_externo}:
    get:
      tags: [Mercado Público (B2G)]
      summary: Detalle de una licitación específica
      parameters:
        - name: codigo_externo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle completo de la licitación.

  /api/v1/mercado-publico/compra-agil:
    get:
      tags: [Mercado Público (B2G)]
      summary: Oportunidades de Compra Ágil (< 300 UTM)
      responses:
        '200':
          description: Listado de compras ágiles activas.

  /api/v1/mercado-publico/ordenes-compra:
    get:
      tags: [Mercado Público (B2G)]
      summary: Listado de órdenes de compra (OCs)
      parameters:
        - name: rut_proveedor
          in: query
          schema:
            type: string
        - name: codigo_organismo
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Listado de OCs.

  /api/v1/mercado-publico/ordenes-compra/{codigo_oc}:
    get:
      tags: [Mercado Público (B2G)]
      summary: Detalle de orden de compra por código
      parameters:
        - name: codigo_oc
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle de la OC.

  /api/v1/mercado-publico/organismos:
    get:
      tags: [Mercado Público (B2G)]
      summary: Directorio de organismos compradores del Estado
      responses:
        '200':
          description: Listado de organismos.

  /api/v1/mercado-publico/organismos/{id}:
    get:
      tags: [Mercado Público (B2G)]
      summary: Perfil de organismo comprador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ficha del organismo.

  /api/v1/mercado-publico/proveedores/{rut}:
    get:
      tags: [Mercado Público (B2G)]
      summary: Perfil B2G 360° de proveedor por RUT
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ficha B2G y score de madurez.

  /api/v1/mercado-publico/proveedores/{rut}/vs-mercado:
    get:
      tags: [Mercado Público (B2G)]
      summary: Benchmark del proveedor vs la industria UNSPSC
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comparativa contra medianas de mercado.

  /api/v1/mercado-publico/proveedores/{rut}/oportunidades:
    get:
      tags: [Mercado Público (B2G)]
      summary: Licitaciones recomendadas para el proveedor
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Licitaciones sugeridas.

  /api/v1/mercado-publico/benchmarks:
    get:
      tags: [Mercado Público (B2G)]
      summary: Benchmarks sectoriales por categoría UNSPSC
      responses:
        '200':
          description: Métricas agregadas por rubro.

  /api/v1/mercado-publico/metricas/{rut}:
    get:
      tags: [Mercado Público (B2G)]
      summary: Métricas consolidadas M1 a M10 por RUT
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Métricas M1-M10.

  # ── Datos Económicos & Macro ──────────────────────────────────────────────
  /api/v1/data/economy:
    get:
      tags: [Datos Económicos & Macro]
      summary: Snapshot macroeconómico consolidado de Chile (UF, IPC, TPM, Dólar)
      responses:
        '200':
          description: Indicadores económicos locales.

  /api/v1/data/macro:
    get:
      tags: [Datos Económicos & Macro]
      summary: Indicadores macroeconómicos globales FRED (Cobre, WTI, Fed Rate, USD/CLP)
      responses:
        '200':
          description: Indicadores FRED globales.

  /api/v1/data/economy/bcch/uf:
    get:
      tags: [Datos Económicos & Macro]
      summary: Valor diario y serie de la UF (Banco Central de Chile)
      responses:
        '200':
          description: Datos UF.

  /api/v1/data/economy/cmf/quiebras-insolvencias:
    get:
      tags: [Datos Económicos & Macro]
      summary: Radar Forense de insolvencias y procedimientos concursales
      responses:
        '200':
          description: Registros de insolvencia.

  # ── Bralidus Intelligence (GraphRAG) ───────────────────────────────────────
  /api/v1/intel/query:
    post:
      tags: [Bralidus Intelligence (GraphRAG)]
      summary: GraphRAG Unificado (Macro + Normativa + B2G)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                top_k:
                  type: integer
                  default: 5
      responses:
        '200':
          description: Respuesta con contexto GraphRAG citable.

  /api/v1/intel/query/moe:
    post:
      tags: [Bralidus Intelligence (GraphRAG)]
      summary: Mixture of Experts (MoE 5 Expertos + Radar Forense)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                max_experts:
                  type: integer
                  default: 3
      responses:
        '200':
          description: Respuesta MoE con expertos activados y frescura de datos.

  /api/v1/intel/entities:
    get:
      tags: [Bralidus Intelligence (GraphRAG)]
      summary: Catálogo de entidades macroeconómicas y normativas indexadas
      responses:
        '200':
          description: Lista de entidades.

  # ── RAG & Vault Vectorial ────────────────────────────────────────────────
  /api/v1/rag/query:
    post:
      tags: [RAG & Vault Vectorial]
      summary: Búsqueda semántica vectorial HNSW (pgvector)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
      responses:
        '200':
          description: Nodos más relevantes por similitud.

  /api/v1/rag/ingest/text:
    post:
      tags: [RAG & Vault Vectorial]
      summary: Ingesta y vectorización de texto plano en espacio privado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [texts]
              properties:
                texts:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Confirmación de ingesta.

  /api/v1/rag/ingest/file:
    post:
      tags: [RAG & Vault Vectorial]
      summary: Carga y parsing automático de documentos (PDF, MD, Docx)
      responses:
        '200':
          description: Documento parsed e ingestado.

  # ── Grafo Societario (S-Pulse) ───────────────────────────────────────────
  /api/v1/data/spulse/companies/search:
    get:
      tags: [Grafo Societario (S-Pulse)]
      summary: Buscador de empresas en el registro societario por nombre o RUT
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Listado de empresas coincidentes.

  /api/v1/data/spulse/companies/{rut}/profile:
    get:
      tags: [Grafo Societario (S-Pulse)]
      summary: Ficha 360° societaria (socios, mallas y alertas de riesgo)
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Perfil 360° societario.

  /api/v1/data/spulse/companies/{rut}/network:
    get:
      tags: [Grafo Societario (S-Pulse)]
      summary: Grafo de redes y mallas de la empresa
      parameters:
        - name: rut
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Nodos y aristas para render UI.

  /api/v1/data/spulse/relationships/{rel_id}/source:
    get:
      tags: [Grafo Societario (S-Pulse)]
      summary: Fuente legal auditable de una relación societaria
      parameters:
        - name: rel_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Documento fuente (Diario Oficial, escritura).

  # ── Webhooks & Servicios ─────────────────────────────────────────────────
  /api/v1/validate:
    post:
      tags: [Webhooks & Servicios]
      summary: Validación de startup (Wizard 18 entregables)
      responses:
        '200':
          description: Resultado de la validación.

  /api/v1/webhooks:
    post:
      tags: [Webhooks & Servicios]
      summary: Registra un webhook para recibir notificaciones en tiempo real
      responses:
        '200':
          description: Webhook registrado.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer